feat: added SMPP property 'systemType'
The systemType parameter is used to categorize the type of ESME that is binding to the SMSC. Examples include “VMS” (voice mail system) and “OTA” (over-the-air activation system).
Specification of the systemType is optional - some SMSC’s may not require ESME’s to provide this detail. In this case, the ESME can set the systemType to NULL.